issue: 3514044 Squash qp_mgr_eth_mlx5 to qp_mgr
